Bugfix: Moved CutProvider out from "positioning" package for backwards compatibility.
