+++ eyes.sdk.core/src/main/java/com/applitools/eyes/FixedCutProvider.java
@@ -1,6 +1,5 @@
 package com.applitools.eyes;
 
-import com.applitools.eyes.positioning.CutProvider;
 import com.applitools.utils.ImageUtils;
 
 import java.awt.image.BufferedImage;
